fix: pypto+simpler CPU simulation and update docs
- Remove pto2_rt_init_tensor_pool from pypto orchestration codegen (simpler
  does not declare it); update pypto submodule
- Map fp64 to FP32 in TypeMapper (PyPTO has no FP64)
- Fix test_phase1_functional: use random init for CPU run test
- Fix test_ttir_converter: unsupported op test, dtype expectations
- Fix test_mlir_parser: correct 2D tensor shape assertion
- Update README with full build env and execution instructions
- Add examples/run_elementwise_e2e.py for TTIR→PyPTO→CPU verification

+++ src/triton_adapter/ttir_converter.py
@@ -60,8 +60,8 @@ class TypeMapper:
         "f16": DataType.FP16,
         "fp32": DataType.FP32,
         "f32": DataType.FP32,
-        "fp64": DataType.FP64,
-        "f64": DataType.FP64,
+        "fp64": DataType.FP32,  # PyPTO 无 FP64，映射到 FP32
+        "f64": DataType.FP32,
     }
 
     def map_dtype(self, ttir_dtype: str) -> DataType:
